Split live service flow and expose execution profiles
