+++ internal/service/live.go
@@ -674,395 +674,6 @@ func (p *Platform) resolveLivePositionStrategyVersionID(accountID, symbol string
 	return ""
 }
 
-func (p *Platform) refreshLiveSessionProtectionState(session domain.LiveSession) (domain.LiveSession, error) {
-	account, err := p.store.GetAccount(session.AccountID)
-	if err != nil {
-		return domain.LiveSession{}, err
-	}
-	snapshot := cloneMetadata(mapValue(account.Metadata["liveSyncSnapshot"]))
-	openOrders := metadataList(snapshot["openOrders"])
-	sessionSymbol := NormalizeSymbol(firstNonEmpty(stringValue(session.State["symbol"]), stringValue(session.State["lastSymbol"])))
-	position, found, err := p.resolvePaperSessionPositionSnapshot(session.AccountID, sessionSymbol)
-	if err != nil {
-		return domain.LiveSession{}, err
-	}
-
-	protectedOrders := make([]map[string]any, 0)
-	stopOrders := make([]map[string]any, 0)
-	takeProfitOrders := make([]map[string]any, 0)
-	for _, item := range openOrders {
-		if sessionSymbol != "" && NormalizeSymbol(stringValue(item["symbol"])) != sessionSymbol {
-			continue
-		}
-		if !isProtectionOrder(item) {
-			continue
-		}
-		protectedOrders = append(protectedOrders, cloneMetadata(item))
-		if isStopProtectionOrder(item) {
-			stopOrders = append(stopOrders, cloneMetadata(item))
-		}
-		if isTakeProfitProtectionOrder(item) {
-			takeProfitOrders = append(takeProfitOrders, cloneMetadata(item))
-		}
-	}
-
-	state := cloneMetadata(session.State)
-	state["recoveredPosition"] = position
-	state["hasRecoveredPosition"] = found
-	state["recoveredProtectionOrders"] = protectedOrders
-	state["recoveredProtectionCount"] = len(protectedOrders)
-	state["recoveredStopOrderCount"] = len(stopOrders)
-	state["recoveredTakeProfitOrderCount"] = len(takeProfitOrders)
-	state["lastProtectionRecoveryAt"] = time.Now().UTC().Format(time.RFC3339)
-	state["lastProtectionRecoverySymbol"] = sessionSymbol
-	state["recoveredStopOrder"] = firstMetadataOrEmpty(stopOrders)
-	state["recoveredTakeProfitOrder"] = firstMetadataOrEmpty(takeProfitOrders)
-
-	status := "flat"
-	switch {
-	case found && len(protectedOrders) > 0:
-		status = "protected-open-position"
-	case found:
-		status = "unprotected-open-position"
-	}
-	state["positionRecoveryStatus"] = status
-	state["protectionRecoveryStatus"] = status
-	if found {
-		appendTimelineEvent(state, "recovery", time.Now().UTC(), "live-position-recovered", map[string]any{
-			"symbol":               sessionSymbol,
-			"protectionCount":      len(protectedOrders),
-			"stopOrderCount":       len(stopOrders),
-			"takeProfitOrderCount": len(takeProfitOrders),
-			"status":               status,
-		})
-	}
-	return p.store.UpdateLiveSessionState(session.ID, state)
-}
-
-func (p *Platform) refreshLiveSessionPositionContext(session domain.LiveSession, eventTime time.Time, source string) (domain.LiveSession, error) {
-	refreshed, err := p.refreshLiveSessionProtectionState(session)
-	if err != nil {
-		return domain.LiveSession{}, err
-	}
-	state := cloneMetadata(refreshed.State)
-	symbol := NormalizeSymbol(firstNonEmpty(stringValue(state["symbol"]), stringValue(state["lastSymbol"])))
-	if symbol == "" {
-		return refreshed, nil
-	}
-	positionSnapshot, foundPosition, err := p.resolvePaperSessionPositionSnapshot(refreshed.AccountID, symbol)
-	if err != nil {
-		return domain.LiveSession{}, err
-	}
-	state["recoveredPosition"] = positionSnapshot
-	state["hasRecoveredPosition"] = foundPosition
-	state["lastRecoveredPositionAt"] = eventTime.UTC().Format(time.RFC3339)
-	state["positionRecoverySource"] = firstNonEmpty(source, "live-position-refresh")
-	if !foundPosition {
-		delete(state, "livePositionState")
-		state["lastLivePositionState"] = map[string]any{}
-		if !boolValue(mapValue(state["virtualPosition"])["virtual"]) {
-			state["positionRecoveryStatus"] = "flat"
-		}
-		return p.store.UpdateLiveSessionState(refreshed.ID, state)
-	}
-
-	version, err := p.resolveCurrentStrategyVersion(refreshed.StrategyID)
-	if err != nil {
-		return domain.LiveSession{}, err
-	}
-	parameters, err := p.resolveLiveSessionParameters(refreshed, version)
-	if err != nil {
-		return domain.LiveSession{}, err
-	}
-	timeframe := firstNonEmpty(stringValue(state["signalTimeframe"]), stringValue(parameters["signalTimeframe"]))
-	signalBarStates := cloneMetadata(mapValue(state["lastStrategyEvaluationSignalBarStates"]))
-	if len(signalBarStates) == 0 {
-		bootstrapStates, bootstrapErr := p.liveSignalBarStates(symbol, timeframe)
-		if bootstrapErr == nil {
-			signalBarStates = bootstrapStates
-			state["lastStrategyEvaluationSignalBarStates"] = signalBarStates
-			state["lastStrategyEvaluationSignalBarBootstrap"] = "market-cache"
-		}
-	}
-	signalBarState, _ := pickSignalBarState(signalBarStates, symbol, timeframe)
-	if signalBarState == nil {
-		return p.store.UpdateLiveSessionState(refreshed.ID, state)
-	}
-	marketPrice := firstPositive(parseFloatValue(positionSnapshot["markPrice"]), parseFloatValue(mapValue(signalBarState["current"])["close"]))
-	livePositionState := evaluateLivePositionState(parameters, positionSnapshot, signalBarState, marketPrice)
-	if len(livePositionState) == 0 {
-		return p.store.UpdateLiveSessionState(refreshed.ID, state)
-	}
-	state["livePositionState"] = livePositionState
-	state["lastLivePositionState"] = livePositionState
-	state["lastPositionContextRefreshAt"] = eventTime.UTC().Format(time.RFC3339)
-	state["lastPositionContextSource"] = firstNonEmpty(source, "live-position-refresh")
-	if boolValue(livePositionState["protected"]) && len(metadataList(state["recoveredProtectionOrders"])) > 0 {
-		state["positionRecoveryStatus"] = "protected-open-position"
-	}
-	return p.store.UpdateLiveSessionState(refreshed.ID, state)
-}
-
-func isProtectionOrder(order map[string]any) bool {
-	orderType := strings.ToUpper(strings.TrimSpace(firstNonEmpty(stringValue(order["origType"]), stringValue(order["type"]))))
-	if boolValue(order["reduceOnly"]) || boolValue(order["closePosition"]) {
-		return true
-	}
-	return strings.Contains(orderType, "STOP") || strings.Contains(orderType, "TAKE_PROFIT")
-}
-
-func isStopProtectionOrder(order map[string]any) bool {
-	orderType := strings.ToUpper(strings.TrimSpace(firstNonEmpty(stringValue(order["origType"]), stringValue(order["type"]))))
-	return strings.Contains(orderType, "STOP")
-}
-
-func isTakeProfitProtectionOrder(order map[string]any) bool {
-	orderType := strings.ToUpper(strings.TrimSpace(firstNonEmpty(stringValue(order["origType"]), stringValue(order["type"]))))
-	return strings.Contains(orderType, "TAKE_PROFIT")
-}
-
-func firstMetadataOrEmpty(items []map[string]any) map[string]any {
-	if len(items) == 0 {
-		return map[string]any{}
-	}
-	return cloneMetadata(items[0])
-}
-
-func (p *Platform) SyncLiveSession(sessionID string) (domain.LiveSession, error) {
-	session, err := p.store.GetLiveSession(sessionID)
-	if err != nil {
-		return domain.LiveSession{}, err
-	}
-	return p.syncLatestLiveSessionOrder(session, time.Now().UTC())
-}
-
-func (p *Platform) syncActiveLiveSessions(eventTime time.Time) error {
-	sessions, err := p.ListLiveSessions()
-	if err != nil {
-		return err
-	}
-	for _, session := range sessions {
-		if !strings.EqualFold(session.Status, "RUNNING") {
-			continue
-		}
-		if strings.TrimSpace(stringValue(session.State["lastDispatchedOrderId"])) == "" {
-			continue
-		}
-		_, _ = p.syncLatestLiveSessionOrder(session, eventTime)
-	}
-	return nil
-}
-
-func (p *Platform) DispatchLiveSessionIntent(sessionID string) (domain.Order, error) {
-	session, err := p.store.GetLiveSession(sessionID)
-	if err != nil {
-		return domain.Order{}, err
-	}
-	return p.dispatchLiveSessionIntent(session)
-}
-
-func (p *Platform) dispatchLiveSessionIntent(session domain.LiveSession) (domain.Order, error) {
-	if !strings.EqualFold(session.Status, "RUNNING") && !strings.EqualFold(session.Status, "READY") {
-		return domain.Order{}, fmt.Errorf("live session %s is not dispatchable in status %s", session.ID, session.Status)
-	}
-
-	proposalMap := cloneMetadata(mapValue(firstNonEmptyMapValue(session.State["lastExecutionProposal"], session.State["lastStrategyIntent"])))
-	if len(proposalMap) == 0 {
-		return domain.Order{}, fmt.Errorf("live session %s has no execution proposal", session.ID)
-	}
-	proposal := executionProposalFromMap(proposalMap)
-	if !strings.EqualFold(proposal.Status, "dispatchable") {
-		return domain.Order{}, fmt.Errorf("live session %s execution proposal is not dispatchable: %s", session.ID, firstNonEmpty(proposal.Status, "unknown"))
-	}
-
-	version, err := p.resolveCurrentStrategyVersion(session.StrategyID)
-	if err != nil {
-		return domain.Order{}, err
-	}
-	order := buildLiveOrderFromExecutionProposal(session, version.ID, proposal, proposalMap)
-	created, createErr := p.CreateOrder(order)
-	if createErr != nil && created.ID == "" {
-		return domain.Order{}, createErr
-	}
-
-	state := cloneMetadata(session.State)
-	intentSignature := buildLiveIntentSignature(proposalMap)
-	dispatchedAt := time.Now().UTC()
-	state["lastDispatchedOrderId"] = created.ID
-	state["lastDispatchedOrderStatus"] = created.Status
-	if isTerminalOrderStatus(created.Status) {
-		state["lastSyncedOrderId"] = created.ID
-		state["lastSyncedOrderStatus"] = created.Status
-	}
-	state["lastDispatchedAt"] = dispatchedAt.Format(time.RFC3339)
-	state["lastDispatchedIntent"] = proposalMap
-	state["lastDispatchedIntentSignature"] = intentSignature
-	delete(state, "lastExecutionTimeoutAt")
-	delete(state, "lastExecutionTimeoutReason")
-	delete(state, "lastExecutionTimeoutIntentSignature")
-	if shouldAdvanceLivePlanForOrderStatus(created.Status) {
-		state["planIndex"] = resolveNextLivePlanIndex(state)
-		state["lastEventTime"] = firstNonEmpty(stringValue(proposalMap["plannedEventAt"]), dispatchedAt.Format(time.RFC3339))
-		state["lastEventSide"] = created.Side
-		state["lastEventRole"] = proposal.Role
-		state["lastEventReason"] = proposal.Reason
-		delete(state, "lastStrategyIntent")
-		delete(state, "lastExecutionProposal")
-	} else {
-		state["lastDispatchRejectedAt"] = dispatchedAt.Format(time.RFC3339)
-		state["lastDispatchRejectedStatus"] = created.Status
-		if shouldMarkLiveExecutionFallback(created) {
-			state["lastExecutionTimeoutAt"] = dispatchedAt.Format(time.RFC3339)
-			state["lastExecutionTimeoutReason"] = "maker-rejected-post-only"
-			state["lastExecutionTimeoutIntentSignature"] = intentSignature
-		}
-	}
-	if createErr != nil {
-		state["lastAutoDispatchError"] = createErr.Error()
-		state["lastAutoDispatchAttemptAt"] = dispatchedAt.Format(time.RFC3339)
-	} else {
-		delete(state, "lastAutoDispatchError")
-	}
-	appendTimelineEvent(state, "order", dispatchedAt, "live-intent-dispatched", map[string]any{
-		"orderId": created.ID,
-		"side":    created.Side,
-		"symbol":  created.Symbol,
-		"price":   created.Price,
-		"role":    proposal.Role,
-		"reason":  proposal.Reason,
-	})
-	if strings.EqualFold(created.Status, "FILLED") {
-		if _, syncErr := p.SyncLiveAccount(session.AccountID); syncErr != nil {
-			state["lastSyncError"] = syncErr.Error()
-		}
-	}
-	updatedSession, _ := p.store.UpdateLiveSessionState(session.ID, state)
-	if strings.EqualFold(created.Status, "FILLED") && updatedSession.ID != "" {
-		if refreshed, refreshErr := p.refreshLiveSessionPositionContext(updatedSession, dispatchedAt, "live-order-fill-sync"); refreshErr == nil {
-			updatedSession = refreshed
-		}
-	}
-	if updatedSession.ID != "" {
-		_, _ = p.syncLatestLiveSessionOrder(updatedSession, time.Now().UTC())
-	}
-	if createErr != nil {
-		return created, createErr
-	}
-	return created, nil
-}
-
-func buildLiveOrderFromExecutionProposal(session domain.LiveSession, strategyVersionID string, proposal ExecutionProposal, proposalMap map[string]any) domain.Order {
-	orderType := strings.ToUpper(strings.TrimSpace(firstNonEmpty(proposal.Type, "MARKET")))
-	quantity := firstPositive(parseFloatValue(session.State["defaultOrderQuantity"]), firstPositive(proposal.Quantity, 0.001))
-	price := proposal.PriceHint
-	if orderType != "MARKET" {
-		price = firstPositive(proposal.LimitPrice, proposal.PriceHint)
-	}
-	return domain.Order{
-		AccountID:         session.AccountID,
-		StrategyVersionID: strategyVersionID,
-		Symbol:            NormalizeSymbol(firstNonEmpty(proposal.Symbol, stringValue(session.State["symbol"]))),
-		Side:              strings.ToUpper(strings.TrimSpace(proposal.Side)),
-		Type:              orderType,
-		Quantity:          quantity,
-		Price:             price,
-		Metadata: map[string]any{
-			"source":             "live-session-intent",
-			"liveSessionId":      session.ID,
-			"signalKind":         proposal.SignalKind,
-			"dispatchMode":       stringValue(session.State["dispatchMode"]),
-			"timeInForce":        proposal.TimeInForce,
-			"postOnly":           proposal.PostOnly,
-			"reduceOnly":         proposal.ReduceOnly,
-			"executionStrategy":  proposal.ExecutionStrategy,
-			"executionExpiresAt": stringValue(proposal.Metadata["executionExpiresAt"]),
-			"executionProposal":  cloneMetadata(proposalMap),
-			"intent":             cloneMetadata(proposalMap),
-		},
-	}
-}
-
-func (p *Platform) applyLiveVirtualInitialEvent(session domain.LiveSession, proposalMap map[string]any, eventTime time.Time) (domain.LiveSession, error) {
-	proposal := executionProposalFromMap(proposalMap)
-	state := cloneMetadata(session.State)
-	intentSignature := buildLiveIntentSignature(proposalMap)
-	entryPrice := firstPositive(
-		parseFloatValue(proposalMap["plannedPrice"]),
-		firstPositive(
-			parseFloatValue(proposalMap["priceHint"]),
-			firstPositive(
-				parseFloatValue(mapValue(proposalMap["metadata"])["bestAsk"]),
-				parseFloatValue(mapValue(proposalMap["metadata"])["bestBid"]),
-			),
-		),
-	)
-	virtualSide := "LONG"
-	if strings.EqualFold(proposal.Side, "SELL") || strings.EqualFold(proposal.Side, "SHORT") {
-		virtualSide = "SHORT"
-	}
-	state["lastDispatchedAt"] = eventTime.UTC().Format(time.RFC3339)
-	state["lastDispatchedIntent"] = cloneMetadata(proposalMap)
-	state["lastDispatchedIntentSignature"] = intentSignature
-	state["lastDispatchedOrderStatus"] = liveOrderStatusVirtualInitial
-	state["lastSyncedOrderStatus"] = liveOrderStatusVirtualInitial
-	state["lastVirtualSignalAt"] = eventTime.UTC().Format(time.RFC3339)
-	state["lastVirtualSignalType"] = "initial"
-	state["virtualPosition"] = map[string]any{
-		"found":      true,
-		"virtual":    true,
-		"symbol":     NormalizeSymbol(proposal.Symbol),
-		"side":       virtualSide,
-		"quantity":   0.0,
-		"entryPrice": entryPrice,
-		"markPrice":  entryPrice,
-		"reason":     proposal.Reason,
-		"recordedAt": eventTime.UTC().Format(time.RFC3339),
-	}
-	state["planIndex"] = resolveNextLivePlanIndex(state)
-	state["lastEventTime"] = firstNonEmpty(stringValue(proposalMap["plannedEventAt"]), eventTime.UTC().Format(time.RFC3339))
-	state["lastEventSide"] = proposal.Side
-	state["lastEventRole"] = proposal.Role
-	state["lastEventReason"] = proposal.Reason
-	delete(state, "lastStrategyIntent")
-	delete(state, "lastExecutionProposal")
-	delete(state, "lastAutoDispatchError")
-	appendTimelineEvent(state, "strategy", eventTime, "live-virtual-initial-recorded", map[string]any{
-		"side":       proposal.Side,
-		"symbol":     proposal.Symbol,
-		"entryPrice": entryPrice,
-		"reason":     proposal.Reason,
-	})
-	return p.store.UpdateLiveSessionState(session.ID, state)
-}
-
-func (p *Platform) applyLiveVirtualExitEvent(session domain.LiveSession, proposalMap map[string]any, eventTime time.Time) (domain.LiveSession, error) {
-	proposal := executionProposalFromMap(proposalMap)
-	state := cloneMetadata(session.State)
-	intentSignature := buildLiveIntentSignature(proposalMap)
-	state["lastDispatchedAt"] = eventTime.UTC().Format(time.RFC3339)
-	state["lastDispatchedIntent"] = cloneMetadata(proposalMap)
-	state["lastDispatchedIntentSignature"] = intentSignature
-	state["lastDispatchedOrderStatus"] = liveOrderStatusVirtualExit
-	state["lastSyncedOrderStatus"] = liveOrderStatusVirtualExit
-	state["lastVirtualSignalAt"] = eventTime.UTC().Format(time.RFC3339)
-	state["lastVirtualSignalType"] = "exit"
-	delete(state, "virtualPosition")
-	state["planIndex"] = resolveNextLivePlanIndex(state)
-	state["lastEventTime"] = firstNonEmpty(stringValue(proposalMap["plannedEventAt"]), eventTime.UTC().Format(time.RFC3339))
-	state["lastEventSide"] = proposal.Side
-	state["lastEventRole"] = proposal.Role
-	state["lastEventReason"] = proposal.Reason
-	delete(state, "lastStrategyIntent")
-	delete(state, "lastExecutionProposal")
-	delete(state, "lastAutoDispatchError")
-	appendTimelineEvent(state, "strategy", eventTime, "live-virtual-exit-recorded", map[string]any{
-		"side":   proposal.Side,
-		"symbol": proposal.Symbol,
-		"reason": proposal.Reason,
-	})
-	return p.store.UpdateLiveSessionState(session.ID, state)
-}
-
 func (p *Platform) StopLiveSession(sessionID string) (domain.LiveSession, error) {
 	session, err := p.store.UpdateLiveSessionStatus(sessionID, "STOPPED")
 	if err != nil {
@@ -1249,22 +860,29 @@ func (p *Platform) evaluateLiveSessionOnSignal(session domain.LiveSession, runti
 		delete(state, "lastExecutionProposalError")
 		executionProposal = executionProposalToMap(proposal)
 		state["lastExecutionProposal"] = executionProposal
+		state["lastExecutionProfile"] = executionProposalSummary(executionProposal)
 		intent = executionProposal
 		state["lastStrategyIntent"] = executionProposal
 		state["lastStrategyIntentSignature"] = buildLiveIntentSignature(executionProposal)
 	} else {
 		delete(state, "lastSignalIntent")
 		delete(state, "lastExecutionProposal")
+		delete(state, "lastExecutionProfile")
 		delete(state, "lastStrategyIntent")
 		delete(state, "lastStrategyIntentSignature")
 	}
 	appendTimelineEvent(state, "strategy", eventTime, "decision", map[string]any{
-		"action":        decision.Action,
-		"reason":        decision.Reason,
-		"decisionState": stringValue(decision.Metadata["decisionState"]),
-		"signalKind":    stringValue(decision.Metadata["signalKind"]),
-		"signalIntent":  cloneMetadata(mapValue(state["lastSignalIntent"])),
-		"intent":        cloneMetadata(intent),
+		"action":            decision.Action,
+		"reason":            decision.Reason,
+		"decisionState":     stringValue(decision.Metadata["decisionState"]),
+		"signalKind":        stringValue(decision.Metadata["signalKind"]),
+		"signalIntent":      cloneMetadata(mapValue(state["lastSignalIntent"])),
+		"intent":            cloneMetadata(intent),
+		"executionStrategy": stringValue(executionProposal["executionStrategy"]),
+		"executionProfile":  stringValue(mapValue(executionProposal["metadata"])["executionProfile"]),
+		"executionMode":     stringValue(mapValue(executionProposal["metadata"])["executionMode"]),
+		"reduceOnly":        boolValue(executionProposal["reduceOnly"]),
+		"fallback":          boolValue(mapValue(executionProposal["metadata"])["fallbackFromTimeout"]),
 	})
 	if executionProposal != nil && strings.EqualFold(stringValue(executionProposal["status"]), "dispatchable") {
 		state["lastStrategyEvaluationStatus"] = "intent-ready"
@@ -1319,145 +937,6 @@ func (p *Platform) evaluateLiveSessionOnSignal(session domain.LiveSession, runti
 		})
 		_, _ = p.store.UpdateLiveSessionState(updatedSession.ID, state)
 		return err
-	}
-	return nil
-}
-
-func (p *Platform) syncLatestLiveSessionOrder(session domain.LiveSession, eventTime time.Time) (domain.LiveSession, error) {
-	orderID := stringValue(session.State["lastDispatchedOrderId"])
-	if strings.TrimSpace(orderID) == "" {
-		return session, nil
-	}
-	order, err := p.GetOrder(orderID)
-	if err != nil {
-		return session, err
-	}
-	state := cloneMetadata(session.State)
-	if isTerminalOrderStatus(order.Status) {
-		state["lastSyncedOrderId"] = order.ID
-		state["lastSyncedOrderStatus"] = order.Status
-		state["lastDispatchedOrderStatus"] = order.Status
-		if strings.EqualFold(order.Status, "FILLED") {
-			_, _ = p.SyncLiveAccount(session.AccountID)
-		}
-		updated, err := p.store.UpdateLiveSessionState(session.ID, state)
-		if err != nil {
-			return domain.LiveSession{}, err
-		}
-		if strings.EqualFold(order.Status, "FILLED") {
-			if refreshed, refreshErr := p.refreshLiveSessionPositionContext(updated, eventTime, "live-order-sync"); refreshErr == nil {
-				return refreshed, nil
-			}
-		}
-		return updated, nil
-	}
-	if shouldCancelLiveOrderForExecutionTimeout(order, eventTime) {
-		cancelledOrder, cancelErr := p.CancelLiveOrder(order.ID)
-		state["lastSyncAttemptAt"] = eventTime.UTC().Format(time.RFC3339)
-		if cancelErr != nil {
-			state["lastSyncError"] = cancelErr.Error()
-			appendTimelineEvent(state, "order", eventTime, "live-order-cancel-error", map[string]any{
-				"orderId": order.ID,
-				"error":   cancelErr.Error(),
-			})
-			updated, updateErr := p.store.UpdateLiveSessionState(session.ID, state)
-			if updateErr != nil {
-				return domain.LiveSession{}, updateErr
-			}
-			return updated, cancelErr
-		}
-		delete(state, "lastSyncError")
-		state["lastSyncedOrderId"] = cancelledOrder.ID
-		state["lastSyncedOrderStatus"] = cancelledOrder.Status
-		state["lastDispatchedOrderStatus"] = cancelledOrder.Status
-		state["lastSyncedAt"] = eventTime.UTC().Format(time.RFC3339)
-		state["lastExecutionTimeoutAt"] = eventTime.UTC().Format(time.RFC3339)
-		state["lastExecutionTimeoutReason"] = "resting-order-expired"
-		timeoutSignature := buildLiveIntentSignature(mapValue(order.Metadata["executionProposal"]))
-		if timeoutSignature == "" {
-			timeoutSignature = buildLiveIntentSignature(mapValue(order.Metadata["intent"]))
-		}
-		if timeoutSignature != "" {
-			state["lastExecutionTimeoutIntentSignature"] = timeoutSignature
-		}
-		appendTimelineEvent(state, "order", eventTime, "live-order-cancelled-timeout", map[string]any{
-			"orderId":     cancelledOrder.ID,
-			"status":      cancelledOrder.Status,
-			"expiredAt":   stringValue(order.Metadata["executionExpiresAt"]),
-			"orderType":   cancelledOrder.Type,
-			"postOnly":    boolValue(order.Metadata["postOnly"]),
-			"timeInForce": stringValue(order.Metadata["timeInForce"]),
-		})
-		return p.store.UpdateLiveSessionState(session.ID, state)
-	}
-	syncedOrder, err := p.SyncLiveOrder(order.ID)
-	state["lastSyncAttemptAt"] = eventTime.UTC().Format(time.RFC3339)
-	if err != nil {
-		state["lastSyncError"] = err.Error()
-		appendTimelineEvent(state, "order", eventTime, "live-order-sync-error", map[string]any{
-			"orderId": order.ID,
-			"error":   err.Error(),
-		})
-		updated, updateErr := p.store.UpdateLiveSessionState(session.ID, state)
-		if updateErr != nil {
-			return domain.LiveSession{}, updateErr
-		}
-		return updated, err
-	}
-	delete(state, "lastSyncError")
-	state["lastSyncedOrderId"] = syncedOrder.ID
-	state["lastSyncedOrderStatus"] = syncedOrder.Status
-	state["lastDispatchedOrderStatus"] = syncedOrder.Status
-	state["lastSyncedAt"] = time.Now().UTC().Format(time.RFC3339)
-	if strings.EqualFold(syncedOrder.Status, "FILLED") {
-		_, _ = p.SyncLiveAccount(session.AccountID)
-	}
-	appendTimelineEvent(state, "order", eventTime, "live-order-synced", map[string]any{
-		"orderId": syncedOrder.ID,
-		"status":  syncedOrder.Status,
-		"price":   syncedOrder.Price,
-	})
-	updated, err := p.store.UpdateLiveSessionState(session.ID, state)
-	if err != nil {
-		return domain.LiveSession{}, err
-	}
-	if strings.EqualFold(syncedOrder.Status, "FILLED") {
-		if refreshed, refreshErr := p.refreshLiveSessionPositionContext(updated, eventTime, "live-order-sync"); refreshErr == nil {
-			return refreshed, nil
-		}
-	}
-	return updated, nil
-}
-
-func isTerminalOrderStatus(status string) bool {
-	switch strings.ToUpper(strings.TrimSpace(status)) {
-	case "FILLED", "CANCELLED", "REJECTED", liveOrderStatusVirtualInitial, liveOrderStatusVirtualExit:
-		return true
-	default:
-		return false
-	}
-}
-
-func shouldCancelLiveOrderForExecutionTimeout(order domain.Order, eventTime time.Time) bool {
-	if isTerminalOrderStatus(order.Status) {
-		return false
-	}
-	expiresAt := parseOptionalRFC3339(stringValue(order.Metadata["executionExpiresAt"]))
-	if expiresAt.IsZero() {
-		return false
-	}
-	return !eventTime.UTC().Before(expiresAt.UTC())
-}
-
-func shouldAdvanceLivePlanForOrderStatus(status string) bool {
-	return !strings.EqualFold(strings.TrimSpace(status), "REJECTED")
-}
-
-func firstNonEmptyMapValue(values ...any) map[string]any {
-	for _, value := range values {
-		if resolved := cloneMetadata(mapValue(value)); len(resolved) > 0 {
-			return resolved
-		}
 	}
 	return nil
 }
